requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
numpy==1.23.5
pinecone-client==3.2.2
langchain~=0.1.20
langchain-community==0.2.0
langchain-pinecone==0.1.1
groq==0.13.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.0.3
sentence-transformers==2.2.2
tiktoken==0.5.1
pyppeteer==2.0.0
googletrans~=4.0.2
langdetect==1.0.9
wikipedia==1.4.0
feedparser==6.0.10
httpx==0.27.2
asyncio==3.4.3
aiohttp==3.9.0
selenium==4.15.0
lxml==4.9.3

playwright~=1.55.0